Move JarUtils into util subpackage
